+++ examples/diez/design-language/src/index.ts
@@ -1,1 +1,0 @@
-export {DesignLanguage} from './DesignLanguage';